ESPiLight::pulseTrainToString: use size_t for length
Pulsetrain length cannot be negative.
Using size_t for size values is probably the best option.

+++ src/ESPiLight.cpp
@@ -361,7 +361,7 @@ static void fire_callback(protocol_t *protocol, ESPiLightCallBack callback) {
              protocol->repeats, deviceId);
 }
 
-String ESPiLight::pulseTrainToString(const uint16_t *codes, int length) {
+String ESPiLight::pulseTrainToString(const uint16_t *codes, size_t length) {
   int i = 0, x = 0, match = 0;
   int diff = 0;
   int plstypes[MAX_PULSE_TYPES];
